==> borrador/06-registered-estand.ipynb <==
### Estandarizamos los datos

import pandas as pd
import numpy as np
import sidetable
import seaborn as sns
import matplotlib.pyplot as plt
from scipy import stats
import math 
from sklearn.preprocessing import MinMaxScaler
import warnings
from sklearn.preprocessing import RobustScaler
warnings.filterwarnings('ignore')
pd.options.display.max_columns = None
import pickle

df= pd.read_csv('../datos/registrados/05-registrados_encoding.csv',index_col = 0)

df.dtypes

numericas = df.select_dtypes(include=np.number)
numericas.drop(['season','holiday','weekday','weathersit','registered', 'yr', "mnth"],axis =1,inplace = True)
numericas.head(2)

robust = RobustScaler()

robust.fit(numericas)

X_robust = robust.transform(numericas)

numericas_robust = pd.DataFrame(X_robust, columns = numericas.columns)
numericas_robust.head(2)

df.drop(['atemp','hum','windspeed'],axis = 1,inplace=True)

numericas_robust.to_csv('../datos/registrados/06-register-numer_robust.csv')

df_total = pd.concat([numericas_robust,df], axis = 1)

df_total.head(2)

df_total.to_csv('../datos/registrados/06-registrados-cod_est.csv')

df_total.to_pickle('../datos/registrados/06-registrados-cod_est.pickle')

with open('../datos/registrados/06_robust_registered.pkl','wb') as fp:
    pickle.dump(robust,fp)
