==> borrador/16-casual-dtree_rforest-estand.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV
import warnings
warnings.filterwarnings('once')
from sklearn.tree import DecisionTreeRegressor
import datetime as dt
import pickle

df = pd.read_csv('../datos/casual/15-casual-cod_est.csv', index_col=0)
df.head(2)

df.dtypes

df.info()

df.head(2)

df.drop(['dteday'], axis = 1, inplace=True)

---

X = df.drop("casual", axis = 1)
y = df["casual"]

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42) 

arbol = DecisionTreeRegressor(random_state =0)
arbol.fit(x_train, y_train)

#fig = plt.figure(figsize = (10,6))
#tree.plot_tree(arbol, feature_names = x_train.columns, filled = True)
#plt.show()

max_features = np.sqrt(len(x_train.columns))
max_features

print(arbol.tree_.max_depth)

param = {"max_depth": [6,8,10,12,14],
        "max_features": [1,2,3],
        "min_samples_split": [10,40,70],
        "min_samples_leaf": [10,40,70]} 

gs = GridSearchCV(
            estimator=DecisionTreeRegressor(), 
            param_grid= param, 
            cv=10, 
            verbose=-1, 
            return_train_score = True, 
            scoring="neg_mean_squared_error")
gs.fit(x_train, y_train)



mejor_modelo = gs.best_estimator_
mejor_modelo

#fig = plt.figure(figsize=(40, 20))
#tree.plot_tree(mejor_modelo, feature_names=x_train.columns, filled=True);

y_pred_test_dt2 = mejor_modelo.predict(x_test)
y_pred_train_dt2 = mejor_modelo.predict(x_train)


def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    
    
    resultados = {'MAE': [mean_absolute_error(y_test, y_test_pred), mean_absolute_error(y_train, y_train_pred)],
                'MSE': [mean_squared_error(y_test, y_test_pred), mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_test_pred)), np.sqrt(mean_squared_error(y_train, y_train_pred))],
                'R2':  [r2_score(y_test, y_test_pred), r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

dt_results1 = metricas(y_test, y_train, y_pred_test_dt2, y_pred_train_dt2, "1_Dtree_casual_est-cod")
dt_results1

Random forest

X = df.drop("casual", axis = 1)
y = df["casual"]

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

param = {"max_depth": [6,8,10,12,14],
        "max_features": [2,3,4,5],
        "min_samples_split": [10,40,70],
        "min_samples_leaf": [10,40,70]} 

gs_rf = GridSearchCV(
            estimator=RandomForestRegressor(), 
            param_grid= param, 
            cv=10, 
            verbose=-1, 
            return_train_score = True, 
            scoring="neg_mean_squared_error") 


gs_rf.fit(x_train, y_train)

bosque = gs_rf.best_estimator_
bosque

'''for arbol in tqdm(bosque.estimators_):
    fig = plt.figure(figsize=(40, 20))
    tree.plot_tree(arbol, feature_names= x_train.columns, filled = True)
    plt.show()'''

y_pred_test_rf = bosque.predict(x_test)
y_pred_train_rf = bosque.predict(x_train)
def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
   
   
    resultados = {'MAE': [mean_absolute_error(y_test, y_test_pred), mean_absolute_error(y_train, y_train_pred)],
                'MSE': [mean_squared_error(y_test, y_test_pred), mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_test_pred)), np.sqrt(mean_squared_error(y_train, y_train_pred))],
                'R2':  [r2_score(y_test, y_test_pred), r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df


dt_results = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "2_RForest_casual_cod_estand")
dt_results

df_todos_resultados = pd.concat([dt_results, dt_results1], axis = 0)
df_todos_resultados


df_todos_resultados.to_csv('../datos/casual/16-casual-metricas-estand.csv')

# Random Forest es la mejor opcion

with open('../datos/casual/16_decision_casual.pkl','wb') as fp:
    pickle.dump(arbol,fp)

with open('../datos/casual/16_random_casual.pkl','wb') as fp:
    pickle.dump(bosque,fp)
